[terraform-destroy]

+++ .github/workflows/executors/gce.sh
@@ -1,8 +1,6 @@
 #!/usr/bin/env bash
 
-#if: "! contains(toJSON(github.event.commits.*.msg), '[skip-ci]')"
-commit_message=$(git log --format=%B -n 1 ${{ github.event.after }})
-echo "${commit_message}"
+destroy=$1
 
 source ${GITHUB_WORKSPACE}/.github/workflows/cicd-definitions.sh
 echo "${WEBIDS_GCP_SA}" > ${GCLOUD_JSON_KEY_PATH}
@@ -21,7 +19,11 @@ if [[ "${hasBucket}" == "true" ]]; then
     cd ${terraform_path}
     terraform init --backend-config="bucket=${GCLOUD_APP_BUCKET_NAME}" --backend-config="prefix=tf-state"
     terraform plan
-    terraform destroy --auto-approve
+    if [[ -z "${destroy}" ]]; then
+        terraform apply --auto-approve
+    else
+        terraform destroy --auto-approve
+    fi
 else
     echo "skip terraform step for now!"
 fi
